start cribbing a new impl of the basic watch service
I think this is okay, but it's based on Oracle's example code provided
for how to use the WatchService.  I think this might lead me to choose
a better path for modifying what's going on in core.  There's not a
whole lot different about this yet, but I was unaware that WatchService
doesn't recursively register sub-directories, insofar as you can tell
when a directory has been created in a watched dir, but anythign inside
that directory will be a mystery to you.