don't use defrecord types from before-ago
I don't think the model I have in place for storing these watches,
paths and events is working out that well for me.  I am cribbing from
the WatchService example code from Oracle and finding that it's simpler
to just track all events and then lookup the fn's to call on them.